Move `GalleryCachePolicy` to its own file
